Don't allocate temporaries
